requirements.txt
fastapi
uvicorn[standard]
pyowletapi
aiohttp
python-dotenv